+++ src/library/DIPS.Mobile.UI/API/Camera/BarcodeScanning/DidFindBarcodeCallback.cs
@@ -1,3 +1,3 @@
 namespace DIPS.Mobile.UI.API.Camera.BarcodeScanning;
 
-public delegate void DidFindBarcodeCallback(Barcode barcode, List<BarcodeObservation> barcodeObservations);
+public delegate void DidFindBarcodeCallback(BarcodeScanResult barcodeScanResult);